Modified code to comply with coding standard
