--- src/main/java/Alfred/Exceptions/InvalidInputException.java
@@ -1,8 +1,0 @@
-package Alfred.Exceptions;
-
-public class InvalidInputException extends AlfredException {
-    static String ERROR_MESSAGE = "Sorry sir, that's a valid command but invalid input. Likely due to wrong format after '(un)mark', 'delete', 'todo', 'deadline' or 'event'.";
-    public InvalidInputException() {
-        super(InvalidInputException.ERROR_MESSAGE);
-    }
-}
