+++ src/main/java/Alfred/Task/Event.java
@@ -1,18 +1,20 @@
-package Alfred.Task;
+package alfred.task;
 
 import java.time.LocalDateTime;
 
 public class Event extends Task {
-  String type = "E";
-  LocalDateTime dateTime;
+    String type = "E";
+    LocalDateTime dateTime;
 
-  public Event(String description, String dateAndTime) {
-    super(description);
-    this.dateTime = LocalDateTime.parse(dateAndTime);;
-  }
+    public Event(String description, String dateAndTime) {
+        super(description);
+        this.dateTime = LocalDateTime.parse(dateAndTime);
+        ;
+    }
 
-  @Override
-  public String toString() {
-    return "[" + this.type + "]" + super.toString() + " (at: " + super.localDateTimeToString(this.dateTime) + ")";
-  }
+    @Override
+    public String toString() {
+        return "[" + this.type + "]" + super.toString() + " (at: " +
+            super.localDateTimeToString(this.dateTime) + ")";
+    }
 }
