+++ src/main/java/Alfred/Task/Task.java
@@ -1,49 +1,50 @@
-package Alfred.Task;
+package alfred.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.FormatStyle;
 
 public abstract class Task {
-  // class constants
-  private static final String COMPLETION_MARK = "X";
-  private static final String INCOMPLETE_MARK = " ";
-  private static final DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofLocalizedDateTime(
-      FormatStyle.MEDIUM, FormatStyle.MEDIUM);
+    // class constants
+    private static final String COMPLETION_MARK = "X";
+    private static final String INCOMPLETE_MARK = " ";
+    private static final DateTimeFormatter dateTimeFormatter =
+        DateTimeFormatter.ofLocalizedDateTime(
+            FormatStyle.MEDIUM, FormatStyle.MEDIUM);
 
 
-  // instance attributes
-  private final String description;
-  private boolean completed;
+    // instance attributes
+    private final String description;
+    private boolean completed;
 
 
-  Task(String description) {
-    this.description = description;
-    this.completed = false;
-  }
-
-  private String markIfComplete() {
-    if (this.completed) {
-      return Task.COMPLETION_MARK;
-    } else {
-      return Task.INCOMPLETE_MARK;
+    Task(String description) {
+        this.description = description;
+        this.completed = false;
     }
-  }
 
-  public void markComplete() {
-    this.completed = true;
-  }
+    private String markIfComplete() {
+        if (this.completed) {
+            return Task.COMPLETION_MARK;
+        } else {
+            return Task.INCOMPLETE_MARK;
+        }
+    }
 
-  public void markIncomplete() {
-    this.completed = false;
-  }
+    public void markComplete() {
+        this.completed = true;
+    }
 
-  public static String localDateTimeToString(LocalDateTime dateTime) {
-    return dateTime.format(Task.dateTimeFormatter);
-  }
+    public void markIncomplete() {
+        this.completed = false;
+    }
 
-  @Override
-  public String toString() {
-    return "[" + this.markIfComplete() + "] " + this.description;
-  }
+    public static String localDateTimeToString(LocalDateTime dateTime) {
+        return dateTime.format(Task.dateTimeFormatter);
+    }
+
+    @Override
+    public String toString() {
+        return "[" + this.markIfComplete() + "] " + this.description;
+    }
 }
