+++ src/main/java/alfred/command/Command.java
@@ -1,0 +1,13 @@
+package alfred.command;
+
+
+import alfred.exceptions.AlfredException;
+import alfred.storage.AlfredStorage;
+import alfred.ui.AlfredUserInterface;
+
+public abstract class Command {
+    public abstract void execute(AlfredUserInterface ui, AlfredStorage storage) throws
+        AlfredException;
+
+    public abstract boolean isExit();
+}
