+++ src/main/java/alfred/exceptions/InvalidCommandException.java
@@ -1,4 +1,4 @@
-package Alfred.Exceptions;
+package alfred.exceptions;
 
 public class InvalidCommandException extends AlfredException {
     static String ERROR_MESSAGE = "Invalid command, sir. I'm not sure what you want me to do.";
