+++ src/main/java/alfred/exceptions/InvalidDateTimeException.java
@@ -1,0 +1,10 @@
+package alfred.exceptions;
+
+public class InvalidDateTimeException extends AlfredException {
+    static String ERROR_MESSAGE =
+        "Invalid date time format, sir. Check that it follows ISO local format, of <YYYY>-<MM>-<DD>T<HH>:<MINMIN>:<SS>";
+
+    public InvalidDateTimeException() {
+        super(InvalidDateTimeException.ERROR_MESSAGE);
+    }
+}
