+++ src/main/java/alfred/exceptions/InvalidIndexException.java
@@ -1,7 +1,8 @@
-package Alfred.Exceptions;
+package alfred.exceptions;
 
 public class InvalidIndexException extends AlfredException {
     static String ERROR_MESSAGE = "Invalid input, sir. Likely due to an invalid list item.";
+
     public InvalidIndexException() {
         super(InvalidIndexException.ERROR_MESSAGE);
     }
