+++ src/main/java/alfred/exceptions/InvalidInputException.java
@@ -1,0 +1,10 @@
+package alfred.exceptions;
+
+public class InvalidInputException extends AlfredException {
+    static String ERROR_MESSAGE =
+        "Sorry sir, that's a valid command but invalid input. Likely due to wrong format after '(un)mark', 'delete', 'todo', 'deadline' or 'event'.";
+
+    public InvalidInputException() {
+        super(InvalidInputException.ERROR_MESSAGE);
+    }
+}
