+++ src/test/java/Alfred/Task/ToDoTest.java
@@ -1,16 +1,16 @@
-package Alfred.Task;
+package alfred.task;
 
-import Alfred.Task.ToDo;
+import alfred.task.ToDo;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class ToDoTest {
 
-  @Test
-  public void todo_valid_toString() {
-    ToDo todo = new ToDo("Expected description");
-    assertEquals(todo.toString(), "[T][ ] Expected description");
-  }
+    @Test
+    public void todo_valid_toString() {
+        ToDo todo = new ToDo("Expected description");
+        assertEquals(todo.toString(), "[T][ ] Expected description");
+    }
 
 }
