+++ src/test/java/alfred/ui/AlfredUserInterfaceTest.java
@@ -1,4 +1,4 @@
-package Alfred.UI;
+package alfred.ui;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
